+++ server/config/config.go
@@ -1,19 +1,36 @@
 package config
 
 import (
-	"github.com/cosmos/cosmos-sdk/server/config"
+	"errors"
+	"fmt"
+
 	"github.com/spf13/viper"
+
+	"github.com/tendermint/tendermint/libs/strings"
+
+	"github.com/cosmos/cosmos-sdk/server/config"
+
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
 const (
 	// DefaultGRPCAddress is the default address the gRPC server binds to.
 	DefaultGRPCAddress = "0.0.0.0:9900"
 
-	// DefaultEVMAddress is the default address the EVM JSON-RPC server binds to.
-	DefaultEVMAddress = "0.0.0.0:8545"
+	// DefaultJSONRPCAddress is the default address the JSON-RPC server binds to.
+	DefaultJSONRPCAddress = "0.0.0.0:8545"
 
-	// DefaultEVMWSAddress is the default address the EVM WebSocket server binds to.
-	DefaultEVMWSAddress = "0.0.0.0:8546"
+	// DefaultJSONRPCWsAddress is the default address the JSON-RPC WebSocket server binds to.
+	DefaultJSONRPCWsAddress = "0.0.0.0:8546"
+
+	// DefaultEVMTracer is the default vm.Tracer type
+	DefaultEVMTracer = "json"
+
+	DefaultGasCap uint64 = 25000000
+)
+
+var (
+	evmTracers = []string{DefaultEVMTracer, "markdown", "struct", "access_list"}
 )
 
 // GetDefaultAPINamespaces returns the default list of JSON-RPC namespaces that should be enabled
@@ -45,8 +62,9 @@ func AppConfig(denom string) (string, interface{}) {
 	}
 
 	customAppConfig := Config{
-		Config: *srvCfg,
-		EVMRPC: *DefaultEVMConfig(),
+		Config:  *srvCfg,
+		EVM:     *DefaultEVMConfig(),
+		JSONRPC: *DefaultJSONRPCConfig(),
 	}
 
 	customAppTemplate := config.DefaultConfigTemplate + DefaultConfigTemplate
@@ -57,33 +75,76 @@ func AppConfig(denom string) (string, interface{}) {
 // DefaultConfig returns server's default configuration.
 func DefaultConfig() *Config {
 	return &Config{
-		Config: *config.DefaultConfig(),
-		EVMRPC: *DefaultEVMConfig(),
+		Config:  *config.DefaultConfig(),
+		EVM:     *DefaultEVMConfig(),
+		JSONRPC: *DefaultJSONRPCConfig(),
 	}
 }
 
-// EVMRPCConfig defines configuration for the EVM RPC server.
-type EVMRPCConfig struct {
-	// RPCAddress defines the HTTP server to listen on
-	RPCAddress string `mapstructure:"address"`
+// EVMConfig defines the application configuration values for the EVM.
+type EVMConfig struct {
+	// Tracer defines vm.Tracer type that the EVM will use if the node is run in
+	// trace mode. Default: 'json'.
+	Tracer string `mapstructure:"tracer"`
+}
+
+// DefaultEVMConfig returns the default EVM configuration
+func DefaultEVMConfig() *EVMConfig {
+	return &EVMConfig{
+		Tracer: DefaultEVMTracer,
+	}
+}
+
+// Validate returns an error if the tracer type is invalid.
+func (c EVMConfig) Validate() error {
+	if !strings.StringInSlice(c.Tracer, evmTracers) {
+		return fmt.Errorf("invalid tracer type %s, available types: %v", c.Tracer, evmTracers)
+	}
+
+	return nil
+}
+
+// JSONRPCConfig defines configuration for the EVM RPC server.
+type JSONRPCConfig struct {
+	// Address defines the HTTP server to listen on
+	Address string `mapstructure:"address"`
 	// WsAddress defines the WebSocket server to listen on
 	WsAddress string `mapstructure:"ws-address"`
 	// API defines a list of JSON-RPC namespaces that should be enabled
 	API []string `mapstructure:"api"`
 	// Enable defines if the EVM RPC server should be enabled.
 	Enable bool `mapstructure:"enable"`
-	// EnableUnsafeCORS defines if CORS should be enabled (unsafe - use it at your own risk)
-	EnableUnsafeCORS bool `mapstructure:"enable-unsafe-cors"`
+	// GasCap is the global gas cap for eth-call variants.
+	GasCap uint64 `mapstructure:"gas-cap"`
 }
 
-// DefaultEVMConfig returns an EVM config with the JSON-RPC API enabled by default
-func DefaultEVMConfig() *EVMRPCConfig {
-	return &EVMRPCConfig{
-		Enable:           true,
-		API:              GetDefaultAPINamespaces(),
-		RPCAddress:       DefaultEVMAddress,
-		WsAddress:        DefaultEVMWSAddress,
-		EnableUnsafeCORS: false,
+// Validate returns an error if the JSON-RPC configuration fields are invalid.
+func (c JSONRPCConfig) Validate() error {
+	if c.Enable && len(c.API) == 0 {
+		return errors.New("cannot enable JSON-RPC without defining any API namespace")
+	}
+
+	// TODO: validate APIs
+	seenAPIs := make(map[string]bool)
+	for _, api := range c.API {
+		if seenAPIs[api] {
+			return fmt.Errorf("repeated API namespace '%s'", api)
+		}
+
+		seenAPIs[api] = true
+	}
+
+	return nil
+}
+
+// DefaultJSONRPCConfig returns an EVM config with the JSON-RPC API enabled by default
+func DefaultJSONRPCConfig() *JSONRPCConfig {
+	return &JSONRPCConfig{
+		Enable:    true,
+		API:       GetDefaultAPINamespaces(),
+		Address:   DefaultJSONRPCAddress,
+		WsAddress: DefaultJSONRPCWsAddress,
+		GasCap:    DefaultGasCap,
 	}
 }
 
@@ -92,7 +153,8 @@ func DefaultEVMConfig() *EVMRPCConfig {
 type Config struct {
 	config.Config
 
-	EVMRPC EVMRPCConfig `mapstructure:"evm-rpc"`
+	EVM     EVMConfig     `mapstructure:"evm"`
+	JSONRPC JSONRPCConfig `mapstructure:"json-rpc"`
 }
 
 // GetConfig returns a fully parsed Config object.
@@ -101,12 +163,37 @@ func GetConfig(v *viper.Viper) Config {
 
 	return Config{
 		Config: cfg,
-		EVMRPC: EVMRPCConfig{
-			Enable:           v.GetBool("evm-rpc.enable"),
-			API:              v.GetStringSlice("evm-rpc.api"),
-			RPCAddress:       v.GetString("evm-rpc.address"),
-			WsAddress:        v.GetString("evm-rpc.ws-address"),
-			EnableUnsafeCORS: v.GetBool("evm-rpc.enable-unsafe-cors"),
+		EVM: EVMConfig{
+			Tracer: v.GetString("evm.tracer"),
+		},
+		JSONRPC: JSONRPCConfig{
+			Enable:    v.GetBool("json-rpc.enable"),
+			API:       v.GetStringSlice("json-rpc.api"),
+			Address:   v.GetString("json-rpc.address"),
+			WsAddress: v.GetString("json-rpc.ws-address"),
+			GasCap:    v.GetUint64("json-rpc.gas-cap"),
 		},
 	}
+}
+
+// ParseConfig retrieves the default environment configuration for the
+// application.
+func ParseConfig(v *viper.Viper) (*Config, error) {
+	conf := DefaultConfig()
+	err := v.Unmarshal(conf)
+
+	return conf, err
+}
+
+// ValidateBasic returns an error any of the application configuration fields are invalid
+func (c Config) ValidateBasic() error {
+	if err := c.EVM.Validate(); err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrAppConfig, "invalid evm config value: %s", err.Error())
+	}
+
+	if err := c.JSONRPC.Validate(); err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrAppConfig, "invalid json-rpc config value: %s", err.Error())
+	}
+
+	return c.Config.ValidateBasic()
 }