+++ types/chain_id.go
@@ -10,10 +10,12 @@ import (
 )
 
 var (
-	regexChainID     = `[a-z]*`
-	regexSeparator   = `-{1}`
-	regexEpoch       = `[1-9][0-9]*`
-	ethermintChainID = regexp.MustCompile(fmt.Sprintf(`^(%s)%s(%s)$`, regexChainID, regexSeparator, regexEpoch))
+	regexChainID         = `[a-z]{1,}`
+	regexEIP155Separator = `_{1}`
+	regexEIP155          = `[1-9][0-9]*`
+	regexEpochSeparator  = `-{1}`
+	regexEpoch           = `[1-9][0-9]*`
+	ethermintChainID     = regexp.MustCompile(fmt.Sprintf(`^(%s)%s(%s)%s(%s)$`, regexChainID, regexEIP155Separator, regexEIP155, regexEpochSeparator, regexEpoch))
 )
 
 // IsValidChainID returns false if the given chain identifier is incorrectly formatted.
@@ -34,8 +36,8 @@ func ParseChainID(chainID string) (*big.Int, error) {
 	}
 
 	matches := ethermintChainID.FindStringSubmatch(chainID)
-	if matches == nil || len(matches) != 3 || matches[1] == "" {
-		return nil, sdkerrors.Wrap(ErrInvalidChainID, chainID)
+	if matches == nil || len(matches) != 4 || matches[1] == "" {
+		return nil, sdkerrors.Wrapf(ErrInvalidChainID, "%s: %v", chainID, matches)
 	}
 
 	// verify that the chain-id entered is a base 10 integer